Add support for subject custom value for email.
Email sharing is currently poorly supported compared to other sharers. The email subject will be set to the item title which can be a rather lengthy string and is almost never appropriate for a subject. To fix this I've added a check for the custom value key "subject" in SHKMail.m. If found, it will be used for the subject of the email, otherwise the subject will fall back to item.title.