Add MBProgressHUD as a submodule.
MBProgressHUD is is now added as a submodule to ShareKit. In
addition the MBProgressHUD.h/.m files are now referenced in the
project. 

Using a submodule (as opposed to simply copying the MBProgressHUD
files into the project) will allow ShareKit to easily incorporate
updates and bug fixes from this and other imported projects.